requirements.txt
anthropic>=0.40.0
neo4j>=5.0.0
chromadb>=0.4.0
llama-index-core>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
PyPDF2>=3.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
streamlit>=1.35.0
sentence-transformers>=3.0.0
tqdm>=4.66.0
pyvis>=0.3.2
